+++ docs/CHANGE_LOG.md
@@ -5,6 +5,7 @@
 #### **Binding Options:**
 - Added a new binding option called "ignoreNullValues", which states if NULL values should be ignored and not rendered (defaults to false).
 - Added a new binding option called "ignoreFunctionValues", which states if FUNCTION values should be ignored and not rendered (defaults to false).
+- Added a new binding option called "reverseArrayValues", which states if the values from an array should be shown in reverse order (defaults to false).
 
 #### **Public Functions:**
 - Added a new public function "refresh()", which refreshes the UI for a specific element.